fabric/requirements.txt
azure-identity
azure-storage-file-datalake
requests
